wallet: move ListAccounts across all backends
